refactor(ssid): extract phase3 site-groups cluster (phase 4/6)
Move Phase 3 site-group orchestrator, ensure/assign coordinators, and 8
pure plan-building helpers into new _ssid_template_phase3.py, wrapped by
_SsidTemplatePhase3Cluster (proxies through _mm to parent state). Parent
retains _create_site_group and _assign_group_sites so historical test
patches (patch.object(_mod, "mistapi", ...)) that rely on the parent's
module namespace continue to intercept mistapi calls.

Parent trimmed 1277 -> 1095 lines. Manager registers the new cluster in
its self._clusters tuple; cluster method resolution uses class-level
hasattr(type(cluster), name) to avoid __getattr__ recursion.

Verification: pytest 241/241 pass, mypy --strict clean, ruff shows 35
F401 (baseline 30 + 5 new phase3 re-exports — accepted pattern).
